Remove sqlite
- Move all source code to main.go
- Restore .goreleaser.yml
