+++ .goreleaser.yml
@@ -1,33 +1,26 @@
 before:
   hooks:
-    - go mod tidy
+    - go mod download
 
-builds:
-  - id: linux
-    ldflags:
-      - -s -w -extldflags=-static
-    binary: ovh-ddns-update
-    goos:
-      - linux
-    goarch:
-      - amd64
-      - 386
-      - arm
-      - arm64
-    goarm:
-      - 5
-      - 6
-      - 7
-
-  - id: macos
-    ldflags:
-      - -s -w
-    binary: ovh-ddns-update
-    goos:
-      - darwin
-    goarch:
-      - amd64
-      - arm64
+build:
+  ldflags:
+    - -s -w
+  binary: ovh-ddns-update
+  goos:
+    - darwin
+    - linux
+  goarch:
+    - amd64
+    - 386
+    - arm
+    - arm64
+  goarm:
+    - 5
+    - 6
+    - 7
+  ignore:
+    - goos: darwin
+      goarch: 386
 
 archives:
   - name_template: '{{ .Binary }}-{{ .Version }}-{{ .Os }}-{{ .Arch }}{{ if .Arm }}v{{ .Arm }}{{ end }}'
