+++ go.mod
@@ -6,14 +6,9 @@ require (
 	github.com/alexflint/go-arg v1.4.2
 	github.com/glendc/go-external-ip v0.1.0
 	github.com/sirupsen/logrus v1.8.1
-	gorm.io/driver/sqlite v1.2.6
-	gorm.io/gorm v1.22.4
 )
 
 require (
 	github.com/alexflint/go-scalar v1.0.0 // indirect
-	github.com/jinzhu/inflection v1.0.0 // indirect
-	github.com/jinzhu/now v1.1.4 // indirect
-	github.com/mattn/go-sqlite3 v1.14.9 // indirect
-	golang.org/x/sys v0.0.0-20211205182925-97ca703d548d // indirect
+	golang.org/x/sys v0.0.0-20191026070338-33540a1f6037 // indirect
 )
