+++ main.go
@@ -5,8 +5,19 @@ import (
 	"github.com/alexflint/go-arg"
 	externalip "github.com/glendc/go-external-ip"
 	log "github.com/sirupsen/logrus"
+	"io/fs"
 	"net/http"
+	"os"
 )
+
+type ArgsType struct {
+	Database    string `default:".ovh-ddns-update.db" help:"database file"`
+	Domain      string `arg:"positional,required" help:"OVH DynHost domain"`
+	OvhId       string `arg:"positional,required" help:"OVH DynHost identifier"`
+	OvhPassword string `arg:"positional,required" help:"OVH DynHost identifier password"`
+}
+
+var Args ArgsType
 
 func main() {
 	arg.MustParse(&Args)
@@ -15,7 +26,6 @@ func main() {
 		FullTimestamp: true,
 	})
 	log.SetLevel(log.InfoLevel)
-	SetupDB()
 	currentIP := GetCurrentIP()
 	previousIP := GetPreviousIP()
 	if currentIP != previousIP {
@@ -35,9 +45,33 @@ func GetCurrentIP() string {
 	return ip.String()
 }
 
+func GetPreviousIP() string {
+	data, err := os.ReadFile(Args.Database)
+	_, ok := err.(*fs.PathError)
+	if ok {
+		log.Warn("Previous IP not found")
+		return "not_found"
+	} else {
+		HandleError(err)
+	}
+	return string(data)
+}
+
 func UpdateDDNS(ip string) {
 	url := fmt.Sprintf("https://%s:%s@www.ovh.com/nic/update?system=dyndns&hostname=%s&myip=%s", Args.OvhId, Args.OvhPassword, Args.Domain, ip)
 	_, err := http.Get(url)
 	HandleError(err)
 	log.Infof("IP updated (%s)", ip)
+}
+
+func SaveCurrentIP(ip string) {
+	err := os.WriteFile(Args.Database, []byte(ip), 0644)
+	HandleError(err)
+}
+
+func HandleError(err error) {
+	if err != nil {
+		log.SetReportCaller(true)
+		log.Fatal(err)
+	}
 }
